Workorder Type is now fixed and working like it does on the client
Code refactore because the changes made from angular 1.2.29 which the client uses and angular 1.3.20 which Orion uses. The way the ng-true-value works on client was changed between these two releases of angular and no longer supports dynamic inputs. Looked online for a solution and this was the best possbile solution with the least amount of html code. More js was needed but it cut down on js variables and html so I chose this